fix: Column 컴포넌트 분리 및 cards 상태값 끌어올리기 수정
