+++ src/components/atoms/column/column.module.scss
@@ -1,0 +1,29 @@
+.column {
+  border-bottom: solid 1px var(--gray30, #eee);
+  padding: 12px;
+
+  @media (min-width: 1024px) {
+    width: 345px;
+    border-right: 1px solid var(--gray30);
+    padding: 20px;
+  }
+}
+
+.defaultArea {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+.titleArea {
+  display: flex;
+  align-items: center;
+}
+.title {
+  margin-left: 8px;
+  margin-right: 12px;
+  font-size: 1.6rem;
+}
+.settingIcon {
+  margin-right: 0;
+  cursor: pointer;
+}
