+++ src/components/molecules/card/Card.tsx
@@ -6,32 +6,30 @@ import ChipTagWithoutX from '@/components/atoms/chipTag/ChipTagWithoutX';
 import instance from '@/api/axios';
 import { format } from 'date-fns';
 import CardDetailModal from '../modals/cardDetailModal/CardDetailModal';
-import { useCardsContext } from '@/components/context/CardsContext';
 
-// interface Card {
-//   id: number;
-//   title: string;
-//   imageUrl: string;
-//   tags: string[];
-//   createdAt: string;
-//   assignee: {
-//     profileImageUrl: string;
-//   };
-// }
+interface Card {
+  id: number;
+  title: string;
+  imageUrl: string;
+  tags: string[];
+  createdAt: string;
+  assignee: {
+    profileImageUrl: string;
+  };
+}
 
 interface CardProps {
   columnId: number;
   columnTitle: string;
+  handleCardsData: (cardsData: []) => void;
 }
 
 export default function Card({
   columnId,
   columnTitle,
   handleCardsData,
-  cards,
 }: CardProps) {
-  // const [cards, setCards] = useState<Card[]>([]);
-  // const { cards, setCards } = useCardsContext();
+  const [cards, setCards] = useState<Card[]>([]);
   const [isDetailModalOpen, setIsDetailModalOpen] = useState(false);
   const [clickedCardId, setClickedCardId] = useState<number | null>(null);
 
@@ -47,8 +45,8 @@ export default function Card({
         },
       );
       const nextCards = res.data.cards;
+      setCards(nextCards);
       handleCardsData(nextCards);
-      // setCards(nextCards);
     } catch (error) {
       console.error('Error fetching cards:', error);
     }
@@ -56,7 +54,6 @@ export default function Card({
 
   useEffect(() => {
     if (columnId !== undefined) {
-      // columnId가 undefined가 아닐 때에만 호 출
       getCards();
     }
   }, [columnId]);
